feat(dashboard): introduce open store flow and screen
This commit introduces the functionality for users to open an existing project/store.

A new "Open Store" screen has been added, accessible via the "Open" button on the home screen. This screen will guide the user through selecting an existing database.

The `MainStoreManager` has been enhanced to handle opening a store by being provided with either the path to the storage directory or a direct path to the database file itself. This provides more flexibility for the user.